requirements.txt
alembic==1.7.7
aniso8601==9.0.1
attrs==22.1.0
bcrypt==4.0.0
click==8.0.4
dataclasses==0.8
Flask==2.0.3
Flask-Bcrypt==1.0.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-Testing==0.8.1
greenlet==1.1.3
importlib-metadata==4.8.3
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==3.2.0
Mako==1.1.6
MarkupSafe==2.0.1
PyJWT==2.4.0
pyrsistent==0.18.0
pytz==2022.4
six==1.16.0
SQLAlchemy==1.4.41
typing_extensions==4.1.1
Werkzeug==2.0.3
zipp==3.6.0
